Added benches for dnstap data parser.
