+++ benches/default.rs
@@ -2,6 +2,7 @@ use criterion::criterion_main;
 
 mod batch;
 mod buffer;
+mod dnstap;
 mod event;
 mod files;
 mod http;
@@ -15,6 +16,7 @@ mod topology;
 criterion_main!(
     batch::benches,
     buffer::benches,
+    dnstap::benches,
     event::benches,
     files::benches,
     http::benches,
